LocalRunOllama/requirements.txt
streamlit
pandas 
numpy
langchain_ollama